Install OPEA application dashboard(s) for ChatQnA + DocSum
This assume "prometheusNamespace" variable value to be correct.

Metrics dashboard is installed when monitoring is enabled, and scaling
dashboard when also HPA autoscaling is enabled.

NOTE: Regardless of which application installs the dashboard(s), they
are identical except for the title in the Grafana dashboards list (and
dashboard internal "uid").  Both dashboards show metrics for _any_ of
the OPEA applications that process streamed tokens with an LLM, not
just for ChatQnA or DocSum.

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>