Helm installed application metric Grafana dashboards (#800)
* Add Helm chart for shared OPEA application metric dashboards

UIDs identifying the dashboards for Grafana, are based on the chart
"fullname" values.

That way, if dashboard configMaps get installed multiple times as
dependencies for different OPEA application charts, Grafana can
differentiate them based on those UIDs.

(At least as long as those application instances are given different
Helm release names, not just installed to separate namespaces.)

Another alternative would have been omitting the dashboard UIDs, to
let Grafana generate them.  However, that would have meant dashboard
URLs changing on every Helm re-deployment, as UID is part of the
Grafana dashboard URL!

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>

* Dashboard double curly braces need to be quoted for Helm

* s/{{err}}/{{ printf "{{err}}" }}/
* s/{{le}}/{{ printf "{{le}}" }}/

Another possibility would be putting the dashboard JSON spec to a
separate file and reading it to Helm configMap template with:
  {{ .Files.Get "dashboards/metrics.json" | toJson | indent 4 }}

But that would require extra scripting to replace dashboard spec
"title" and "uid" fields with suitable templated values.

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>

* Install OPEA application dashboard(s) for ChatQnA + DocSum

This assume "prometheusNamespace" variable value to be correct.

Metrics dashboard is installed when monitoring is enabled, and scaling
dashboard when also HPA autoscaling is enabled.

NOTE: Regardless of which application installs the dashboard(s), they
are identical except for the title in the Grafana dashboards list (and
dashboard internal "uid").  Both dashboards show metrics for _any_ of
the OPEA applications that process streamed tokens with an LLM, not
just for ChatQnA or DocSum.

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>

* Update monitoring documentation

* Describe new Helm functionality in monitoring.md
* Update kubernetes-addons/Observability/
  - Add index to README + move dashboard import info to its own section
  - Remove duplicate Protheus / Grafana install info
  - Move related files to better locations
* New dashboad screenshots

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>

---------

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>