Change vLLM custom metric from queue size to active requests sum
TGI and TEI queue size metric is only for requests that are waiting to
be processed, but that number can fluctuate a lot, as it's non-zero
only when pod is fully utilized, and needs to buffer requests.  On the
plus side, it's agnostic to how fast engine instance can process the
queries for given model.

vLLM provides also gauge metric for how many requests are currently
being processed (running).  Adding that to the waiting requests count
(queue size) makes the resulting metric much more stable, and allows
scaling up extra replicas before current ones are full. KubeAI autoscaling
is also based on number of active requests, so results will be more
comparable.

Howeover, this means that the suitable threshold will be model and
engine config specific (depending on how request batches HW can run in
parallel).

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>