Rework HPA scaling metrics and their configuration (#1090)
* Disable vLLM warmup with HPA autoscaling

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>

* Simplify HPA rules + add HPA queue size target variable

* Use .Series, .GroupBy and .LabelMatchers to simplify rules
* Drop request latency metric for TGI/TEI.  It depending on
  number of generated tokens, makes it unsuitable as generic metric
  - With that, support for HPA Value type could also be dropped
    (leaving only queue size AverageValue)
* Because vLLM token mean latency metric does not react much
  to vLLM load, and for consistency with TGI/TEI, switch vLLM
  also to be scaled based on queue size
  - KubeAI scales vLLM also based on queue size
* Add queue size target Helm variables for all inferencing engines

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>

* Fix metrics dashboad next tokens rate query for TGI

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>

* Change vLLM custom metric from queue size to active requests sum

TGI and TEI queue size metric is only for requests that are waiting to
be processed, but that number can fluctuate a lot, as it's non-zero
only when pod is fully utilized, and needs to buffer requests.  On the
plus side, it's agnostic to how fast engine instance can process the
queries for given model.

vLLM provides also gauge metric for how many requests are currently
being processed (running).  Adding that to the waiting requests count
(queue size) makes the resulting metric much more stable, and allows
scaling up extra replicas before current ones are full. KubeAI autoscaling
is also based on number of active requests, so results will be more
comparable.

Howeover, this means that the suitable threshold will be model and
engine config specific (depending on how request batches HW can run in
parallel).

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>

* Add autoscaling metric documentation

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>

* Improve ChatQnA HPA enabling comments

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>

* HPA document improvement suggestions from Markus

And use same formatting for all notes.

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>

---------

Signed-off-by: Eero Tamminen <eero.t.tamminen@intel.com>